Moved rtl_tester into a tester util file. Updated tests to use appropriate new names and new family closures
