+++ tests/test_micro.py
@@ -1,130 +1,24 @@
-from collections import namedtuple
 import lassen.asm as asm
-from lassen.sim import gen_pe, gen_pe_type_family
-from lassen.mode import gen_mode_type
-from lassen.isa import DATAWIDTH, gen_alu_type
-from hwtypes import SIntVector, UIntVector, BitVector, Bit, FPVector, RoundingMode
+from lassen import PE_fc, Inst_fc
+from lassen.common import DATAWIDTH, BFloat16_fc
 from lassen.utils import float2bfbin, bfbin2float
+from hwtypes import SIntVector, UIntVector, BitVector, Bit
+
+from collections import namedtuple
+
+import random
 import pytest
-import math
-import random
-import magma
-import peak
-import fault
-import os
-import random
-import shutil
-from peak.assembler import Assembler
 
-class HashableDict(dict):
-    def __hash__(self):
-        return hash(tuple(sorted(self.keys())))
-
-Bit = Bit
-Data = BitVector[DATAWIDTH]
-BFloat16 = FPVector[8, 7,RoundingMode.RNE,False]
-
-
-PE = gen_pe(BitVector.get_family())
+PE = PE_fc(Bit.get_family())
 pe = PE()
-sim_family = gen_pe_type_family(BitVector.get_family())
-Mode = gen_mode_type(sim_family)
+Data = BitVector[DATAWIDTH]
+BFloat16 = BFloat16_fc(Bit.get_family())
 
-# create these variables in global space so that we can reuse them easily
-inst_name = 'inst'
-inst_type = PE.input_t.field_dict[inst_name]
+#TODO reenable RTL tests
+def rtl_tester(*args,**kwargs):
+    pass
 
-_assembler = Assembler(inst_type)
-assembler = _assembler.assemble
-disassembler = _assembler.disassemble
-width = _assembler.width
-layout = _assembler.layout
-pe_magma = gen_pe(magma.get_family(), use_assembler=True)
-instr_magma_type = type(pe_magma.interface.ports[inst_name])
-pe_circuit = peak.wrap_with_disassembler(pe_magma, disassembler, width,
-                                         HashableDict(layout),
-                                         instr_magma_type)
-tester = fault.Tester(pe_circuit, clock=pe_circuit.CLK)
-test_dir = "tests/build"
-
-# Explicitly load `float_DW` lib so we get technology specific mapping with
-# We reset the context because tests/test_pe.py calls compile and pollutes the
-# coreir context causing a "redefinition of module" error
-magma.backend.coreir_.CoreIRContextSingleton().reset_instance()
-magma.compile(f"{test_dir}/WrappedPE", pe_circuit, output="coreir-verilog",
-              coreir_libs={"float_DW"})
-
-# check if we need to use ncsim + cw IP
-cw_dir = "/cad/synopsys/dc_shell/J-2014.09-SP3/dw/sim_ver/"   # noqa
-CAD_ENV = shutil.which("ncsim") and os.path.isdir(cw_dir)
-
-
-def copy_file(src_filename, dst_filename, override=False):
-    if not override and os.path.isfile(dst_filename):
-        return
-    shutil.copy(src_filename, dst_filename)
-
-
-def rtl_tester(test_op, data0=None, data1=None, bit0=None, bit1=None, bit2=None,
-               res=None, res_p=None, clk_en=1, delay=0,
-               data0_delay_values=None, data1_delay_values=None):
-    tester.clear()
-    if hasattr(test_op, "inst"):
-        tester.circuit.inst = assembler(test_op.inst)
-    else:
-        tester.circuit.inst = assembler(test_op)
-    tester.circuit.CLK = 0
-    tester.circuit.clk_en = clk_en
-    if data0 is not None:
-        data0 = BitVector[16](data0)
-        tester.circuit.data0 = data0
-    if data1 is not None:
-        data1 = BitVector[16](data1)
-        tester.circuit.data1 = data1
-    if bit0 is not None:
-        tester.circuit.bit0 = Bit(bit0)
-    if bit1 is not None:
-        tester.circuit.bit1 = Bit(bit1)
-    if bit2 is not None:
-        tester.circuit.bit2 = Bit(bit2)
-    tester.eval()
-
-    for i in range(delay):
-        tester.step(2)
-        if data0_delay_values is not None:
-            tester.circuit.data0 = data0_delay_values[i]
-        if data1_delay_values is not None:
-            tester.circuit.data1 = data1_delay_values[i]
-
-    if res is not None:
-        tester.circuit.O0.expect(res)
-    if res_p is not None:
-        tester.circuit.O1.expect(res_p)
-    if CAD_ENV:
-        # use ncsim
-        libs = ["DW_fp_mult.v", "DW_fp_add.v", "DW_fp_addsub.v"]
-        for filename in libs:
-            copy_file(os.path.join(cw_dir, filename),
-                      os.path.join(test_dir, filename))
-        tester.compile_and_run(target="system-verilog", simulator="ncsim",
-                               directory="tests/build/",
-                               include_verilog_libraries=libs,
-                               skip_compile=True)
-    else:
-        libs = ["DW_fp_mult.v", "DW_fp_add.v"]
-        for filename in libs:
-            copy_file(os.path.join("stubs", filename),
-                      os.path.join(test_dir, filename))
-        # detect if the PE circuit has been built
-        skip_verilator = os.path.isfile(os.path.join(test_dir, "obj_dir",
-                                                     "VWrappedPE__ALL.a"))
-        tester.compile_and_run(target="verilator",
-                               directory=test_dir,
-                               flags=['-Wno-UNUSED', '-Wno-PINNOCONNECT'],
-                               skip_compile=True,
-                               skip_verilator=skip_verilator)
-
-
+CAD_ENV = True
 
 op = namedtuple("op", ["inst", "func"])
 NTESTS = 16
